backend/requirements.txt
django>=5.2.0,<6.0.0
djangorestframework>=3.14.0
channels>=4.0.0
daphne>=4.0.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99
langdetect>=1.0.9
PyJWT>=2.8.0